# backend/openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.1
  title: Anki Clone Application API
  description: This is the API of the Anki Clone Application. It allows you to manage users, learning packages, learning facts, tags and statistics.
paths:
...
  /api/tag:
    get:
      summary: Get all tags
      operationId: getTag
      tags:
        - Tags
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new tag
      operationId: createTag
      tags:
        - Tags
      requestBody:
        description: The tag to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                englishKeyword:
                  type: string
                  description: The english keyword of the tag
                  example: TypeScript
                  required: true
                frenchTranslation:
                  type: string
                  description: The french translation of the tag
                  example: TypeScript
                  required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad Request
        '409':
          description: Conflict
        '500':
          description: Internal Server Error
  /api/tag/{tag_id}:
    put:
      summary: Update a tag by id
      operationId: updateTagById
      tags:
        - Tags
      parameters:
        - name: tag_id
          in: path
          description: The id of the tag
          required: true
          schema:
            type: number
      requestBody:
        description: The tag to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                englishKeyword:
                  type: string
                  description: The english keyword of the tag
                  example: TypeScript
                frenchTranslation:
                  type: string
                  description: The french translation of the tag
                  example: TypeScript
      responses:
        '204':
          description: Updated
        '404':
          description: Not Found
        '409':
          description: Conflict
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a tag by id
      operationId: deleteTagById
      tags:
        - Tags
      parameters:
        - name: tag_id
          in: path
          description: The id of the tag
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Deleted
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # STATISTICS API
  /api/statistics/user/{user_id}:
    get:
      summary: Get the statistics of a user by id
      operationId: getStatisticsByUserId
      tags:
        - Statistics
      parameters:
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: A list of user statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nextDays:
                      type: object
                      additionalProperties:
                        type: integer
                    todayFacts:
                      type: integer
                    globalProgress:
                      type: number
                      format: double
                    nbCreatedPackages:
                      type: integer
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE <> FACT API
...
  /api/package/{package_id}/tag:
    get:
      summary: Get all tags by learning package id
      operationId: getTagByLearningPackageId
      tags:
        - Learning Packages <> Tags
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new tag
      operationId: createTag
      tags:
        - Learning Packages <> Tags
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      requestBody:
        description: The tag to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tagId:
                  type: number
                  description: The id of the tag
                  example: 1
                  required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPackageTag'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/package/{package_id}/tag/{tag_id}:
    delete:
      summary: Delete a tag by learning package id and tag id
      operationId: deleteTagByLearningPackageIdAndTagId
      tags:
        - Learning Packages <> Tags
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
        - name: tag_id
          in: path
          description: The id of the tag
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Deleted
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE <> USER API
...
servers:
  [
    {
      url: http://localhost:3000,
      description: Development server
    }
  ]
components:
  schemas:
...
    Tag:
      type: object
      properties:
        tagId:
          type: number
          description: The id of the tag
          example: 1
        englishKeyword:
          type: string
          description: The english keyword of the tag
          example: TypeScript
        frenchTranslation:
          type: string
          description: The french translation of the tag
          example: TypeScript
      required:
        - tagId
        - englishKeyword
...
    LearningPackageTag:
      type: object
      properties:
        learningPackageTagId:
          type: number
          description: The id of the learning package tag
          example: 1
        packageId:
          type: number
          description: The id of the learning package
          example: 1
        tagId:
          type: number
          description: The id of the tag
          example: 1
      required:
        - learningPackageTagId
        - packageId
        - tagId